requirements.txt
aws-cdk-lib>=2.178.2
constructs>=10.0.0,<11.0.0
boto3>=1.36.21
fastapi[all]>=0.115.8
pydantic>=2.10.6
python-dotenv>=1.0.1
ruff>=0.9.6
mangum>=0.19.0